build: fix linting issues reported by running golangci-lint with different GOOS
