+++ backend/azureblob/azureblob_unsupported.go
@@ -3,4 +3,5 @@
 
 //go:build plan9 || solaris || js
 
+// Package azureblob provides an interface to the Microsoft Azure blob object storage system
 package azureblob
