+++ backend/azurefiles/azurefiles_unsupported.go
@@ -3,4 +3,5 @@
 
 //go:build plan9 || js
 
+// Package azurefiles provides an interface to Microsoft Azure Files
 package azurefiles
