+++ backend/cache/cache_unsupported.go
@@ -3,4 +3,5 @@
 
 //go:build plan9 || js
 
+// Package cache implements a virtual provider to cache existing remotes.
 package cache
