+++ backend/hdfs/hdfs_unsupported.go
@@ -3,4 +3,5 @@
 
 //go:build plan9
 
+// Package hdfs provides an interface to the HDFS storage system.
 package hdfs
