+++ backend/local/xattr_unsupported.go
@@ -1,6 +1,7 @@
+// The pkg/xattr module doesn't compile for openbsd or plan9
+
 //go:build openbsd || plan9
 
-// The pkg/xattr module doesn't compile for openbsd or plan9
 package local
 
 import "github.com/rclone/rclone/fs"
