+++ backend/oracleobjectstorage/oracleobjectstorage_unsupported.go
@@ -3,4 +3,5 @@
 
 //go:build plan9 || solaris || js
 
+// Package oracleobjectstorage provides an interface to the OCI object storage system.
 package oracleobjectstorage
