+++ backend/qingstor/qingstor_unsupported.go
@@ -3,4 +3,6 @@
 
 //go:build plan9 || js
 
+// Package qingstor provides an interface to QingStor object storage
+// Home: https://www.qingcloud.com/
 package qingstor
