+++ backend/storj/storj_unsupported.go
@@ -1,3 +1,4 @@
 //go:build plan9
 
+// Package storj provides an interface to Storj decentralized object storage.
 package storj
