+++ cmd/cachestats/cachestats_unsupported.go
@@ -3,4 +3,5 @@
 
 //go:build plan9 || js
 
+// Package cachestats provides the cachestats command.
 package cachestats
