+++ cmd/ncdu/ncdu_unsupported.go
@@ -3,4 +3,5 @@
 
 //go:build plan9 || js
 
+// Package ncdu implements a text based user interface for exploring a remote
 package ncdu
