+++ fs/fserrors/enospc_error_notsupported.go
@@ -2,7 +2,7 @@
 
 package fserrors
 
-// IsErrNoSpace() on plan9 returns false because
+// IsErrNoSpace on plan9 returns false because
 // plan9 does not support syscall.ENOSPC error.
 func IsErrNoSpace(cause error) (isNoSpc bool) {
 	isNoSpc = false
