+++ fs/logger/logger_test.go
@@ -1,3 +1,6 @@
+//go:build !plan9
+// +build !plan9
+
 package logger_test
 
 import (
